+++ src/test/java/com/openggf/audio/TestAudioManagerRuntimeInstallation.java
@@ -226,6 +226,7 @@ class TestAudioManagerRuntimeInstallation {
                 "beginReversePresentation",
                 "endReversePresentation",
                 "setReversePlaybackRate",
+                "setForwardPlaybackRate",
                 "setRewindHistoryArmed",
                 "captureLogicalSnapshot",
                 "restoreLogicalSnapshot",